backend/requirements.txt
absl-py==0.8.1
astor==0.8.0
Click==7.0
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.3
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.10.0
pyAttributes==0.2.6
python-docx==0.8.10
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
SphinxExtensions==0.1.3
tensorboard==2.0.0
tensorflow-estimator==2.0.1
#tensorflow-gpu==2.0.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
mysql-connector-python

spacy==3.0
flask-cors
minio
jsonlines
gunicorn
clearml
boto3
ipdb
